feat: refactors agent module imports and updates TypeScript configuration
