+++ packages/agent/src/data-generator/index.ts
@@ -1,0 +1,3 @@
+export * from './data-generator.module';
+export * from './data-generator.service';
+export * from './data-repository';
