+++ packages/agent/src/deploy/index.ts
@@ -1,0 +1,5 @@
+export * from './deploy-vercel.dto';
+export * from './deploy.controller';
+export * from './deploy.module';
+export * from './deploy.types';
+export * from './vercel.service';
