+++ packages/agent/src/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './directory.entity';
+export * from './user.entity';
