+++ packages/agent/src/git/index.ts
@@ -1,0 +1,3 @@
+export * from './git.module';
+export * from './git.provider';
+export * from './github.service';
