+++ packages/agent/src/http/index.ts
@@ -1,0 +1,3 @@
+export * from './agent-http.controller';
+export * from './agent-http.module';
+export * from './agent.service';
