--- packages/agent/src/index.ts
@@ -1,48 +1,0 @@
-// Agent HTTP
-export * from './http/agent-http.module';
-export * from './http/agent-http.controller';
-export * from './http/agent.service';
-
-// Data Generator
-export * from './data-generator/data-generator.module';
-export * from './data-generator/data-generator.service';
-
-// Items Generator
-export * from './items-generator/items-generator.module';
-export * from './items-generator/items-generator.service';
-export * from './items-generator/item-submission.service';
-
-// Markdown Generator
-export * from './markdown-generator/markdown-generator.module';
-export * from './markdown-generator/markdown-generator.service';
-
-// Website Generator
-export * from './website-generator/website-generator.module';
-export * from './website-generator/website-generator.service';
-export * from './website-generator/website-update.service';
-
-// AI Service
-export * from './ai';
-
-// Git
-export * from './git/git.module';
-export * from './git/github.service';
-
-// Database
-export * from './database/database.module';
-export * from './database/database.config';
-export * from './database/database-config.factory';
-export * from './database/directory.repository';
-
-// Entities
-export * from './entities/directory.entity';
-export * from './entities/user.entity';
-
-// DTOs
-export * from './dto/create-directory.dto';
-export * from './items-generator/dto';
-export * from './website-generator/dto/update-website-repository.dto';
-
-// Deploy
-export * from './deploy/deploy.module';
-export * from './deploy/vercel.service';
