+++ pnpm-workspace.yaml
@@ -3,9 +3,10 @@ packages:
     - packages/*
 
 onlyBuiltDependencies:
-    - better-sqlite3
-    - core-js
-    - msgpackr-extract
-    - sqlite3
     - '@nestjs/core'
     - '@swc/core'
+    - better-sqlite3
+    - core-js
+    - esbuild
+    - msgpackr-extract
+    - sqlite3
